+++ src/tepkit/__main__.py
@@ -1,0 +1,4 @@
+from tepkit.cli.typer.entry import main
+
+if __name__ == "__main__":
+    main()