+++ src/tepkit/functions/mlip/__init__.py
@@ -1,0 +1,1 @@
+from .cfg_slice import cfg_slice_cli