+++ ChangeLog.md
@@ -1,6 +1,31 @@
 # Change Log
 
-## 0.2.1 (2025-10-11)
+## 0.2.2 (2026)
+
+### CLI Changes
+
+- **Update:** Add 1 new option `--list/-l` into `vasp check-dirs` command to activate list mode.
+
+### Core Changes
+
+- None
+
+### Bug Fixes
+
+- **Fixed:** Option `--group/-g` not working in `vasp check-dirs` command.
+- **Fixed:** Different behavior of `*` on windows and linux caused by `windows_expand_args`.
+- **Fixed:** Incorrect auto name when saving CSV file in `tepkit.functions.phonopy.rms.rms()`.  
+  (now `tepkit.RMS_of_3rdIFCs.csv` -> `tepkit.RMS_of_2ndIFCs.csv`)
+- **Fixed:** Support `lib64` in `Fourthorder_tools/setup.py` file.
+- 
+### Miscellaneous
+
+- Add the `__main__.py` file to enable `python -m tepkit` entry.
+- Update document.
+
+---
+
+## 0.2.1 (2025-12-29)
 
 ### CLI Changes